Fixed imports and hash implementation.
The hash implementation looked like it was written for python 2.X. In python 3.X the strings are utf and not binary, and thus the concatenation and then hashing failed. Changed by making the prefix bytes binary strings.

Added local.py which is a simple harness to show that this is somewhat working. None of this code contains the ability to generate proofs, so we need to wait until we get one from a Trillian server to check that the verification is working.